% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data_and_temp_folder.R
\name{reduce_data_and_temp_folder}
\alias{reduce_data_and_temp_folder}
\title{reduce_data_and_temp_folder}
\usage{
reduce_data_and_temp_folder(data)
}
\description{
reduce_data_and_temp_folder
}
